requirements.txt
pandas
sqlalchemy
psycopg2-binary
kafka-python
numpy